requirements.txt
streamlit==1.39.0
tensorflow==2.15.0
numpy
pillow